backend_api/uploads/requirements.txt
Flask==2.3.3
tensorflow==2.13.0
Pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
gunicorn==21.2.0
werkzeug==2.3.7
protobuf==4.25.3